Eliminate the redundant get_indices function (closes: gh-195)

+++ tables/vlarray.py
@@ -822,7 +822,7 @@ class VLArray(hdf5extension.VLArray, Leaf):
         """Read rows specified in `coords`."""
         rows = []
         for coord in coords:
-            rows.append(self.read(coord)[0])
+            rows.append(self.read(long(coord))[0])
         return rows
 
     def _g_copy_with_stats(self, group, name, start, stop, step,
